requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-taggit==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
whitenoise==6.0.0
